@@ -1,13 +1,6 @@
-class ApiController < ActionController::Base
+class ApiController < ApplicationController
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :null_session
-  before_filter :set_constants
-
-  # used in application.html.erb to remove sidebar from layout
-  def set_constants
-    @controller = params[:controller]
-    @ability = Ability.new(current_user)
-  end
 
 end
