@@ -1,1 +1,1 @@
-PaperTrail.config.track_associations = false
+#PaperTrail.config.track_associations = false
